==> 00_obtaining_data.ipynb <==
# Project 3: Reddit Post Classification

<i>Pulling information and classifying posts via Pushshift's API</i>

**Author: Brendan McDonnell**

## Problem Statement:

Use Reddit's API to collect posts from two subreddits, r/The_Donald and r/Republican. Using natural language processing (NLP), I will train a classification model to predict which subreddit a given post came from.

**Reddit pages chosen**
- https://www.reddit.com/r/Republican/
- https://www.reddit.com/r/The_Donald/

**From r/Republican Community Details**

/r/Republican is a partisan subreddit. This is a place for Republicans to discuss issues with other Republicans.

**From r/The_Donald Community Details**

The_Donald is a never-ending rally dedicated to the 45th President of the United States, Donald J. Trump.

These two subreddits bear some similarities and should have some overlap. However, it's important to note that r/The_Donald has been quarantined recently for attempting to incite violence against police and other public officials, a violation of the reddit's content policy. It is a well known and used alt-right subreddit. By contrast, r/Republican is a subreddit that should reflect contemporary Republican views. There will likely be some NSFW material on r/The_Donald, but I believe a model like this could be very useful for understanding more about the current Republican viewpoint.

## Step 1: Obtaining The Data via Reddit's API

Using Reddit's API in conjuction with the praw wrapper, I will create csv's of 15,000 rows of data for each of the subreddits chosen.

## Relative Links
- [Pulling the Data](#Pulling-the-Data)
- [Export the Data to CSV](#Export-the-Data-to-CSV)

## Pulling the Data

import praw
import requests
import time
from datetime import datetime
from psaw import PushshiftAPI
import pandas as pd

reddit = praw.Reddit(client_id = 'EvtTIf6Mz_xTUA',     #this is that personal use script key
                     client_secret = '-HelFVGKjgSDv489RnL_0iFmzAM', #this is that secret key
                     user_agent = 'the_donald_v_republican',    #Whatever name you gave your application
                     username = 'B_Mac206',      #your Reddit user name
                     password = 'kbrepuscankt')      #your Reddit password

api = PushshiftAPI(reddit)

for i in api.search_submissions(subreddit='Republica', limit=10):
    print(i.author)

def subreddit_dict(subreddit, limit):
    '''
    Take in subreddit, and spit out df of reddit data for subreddit w/ specified rows (limit).    
    '''
#     Using Heather's code from CHI-FLEX-8 to create for loop to produce df of reddit data
    counter = 0
    sleep_counter = 0
    topics_dict = { "title":[],
                "score":[],
                "id":[],
                "url":[], 
                "comms_num": [],
                "created": [],
                "body":[],
                "user": []}

#     Use a for loop to take the posts gathered by the wrapper and place them into a dictionary.

    for item in api.search_submissions(subreddit=subreddit,limit=limit):
        topics_dict["title"].append(item.title)
        topics_dict["score"].append(item.score)
        topics_dict["id"].append(item.id)
        topics_dict["url"].append(item.url)
        topics_dict["comms_num"].append(item.num_comments)
        topics_dict["created"].append(item.created)
        topics_dict["body"].append(item.selftext)
        topics_dict["user"].append(item.author)
        counter +=1
        sleep_counter += 1
        if sleep_counter > 100:
            print(counter)
            time.sleep(3)
            sleep_counter = 0
    df =  pd.DataFrame(topics_dict)
    df['created'] = df['created'].apply(lambda x: datetime.fromtimestamp(x))
    return df    

# the_donald_df = subreddit_dict('The_Donald', 15_000)

# republican_df = subreddit_dict('Republican', 15_000)

republican_df.shape, the_donald_df.shape

## Export the Data to CSV 

# republican_df.to_csv('data/republican.csv')

# the_donald_df.to_csv('data/the_donald.csv')
