==> 02_EDA_2.ipynb <==
# Project 3: Reddit Post Classification

<i>Pulling information and classifying posts via Pushshift's API</i>

**Author: Brendan McDonnell**

## Step 2: EDA Part 2

Exploring, visualizing, and pulling information from the two datasets before modeling.

## Relative Links
- [Importing Libraries and Datasets Needed](#Importing-Libraries-and-Datasets-Needed)
- [Adding Textblob Sentiment Features](#Adding-Textblob-Sentiment-Features)
- [Further Exploration](#Further-Exploration)
- [Export the Data to CSV](#Export-the-Data-to-CSV)

## Importing Libraries and Datasets Needed

import pandas as pd
import numpy as np
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer
from textblob import TextBlob, Word, Blobber 
import matplotlib.pyplot as plt
import plotly.plotly as py
import plotly.graph_objs as go
import cufflinks as cf

df = pd.read_csv('data/data_comb_w_sent.csv')

df.columns

df[df.duplicated()]

df.drop_duplicates(inplace=True)

## Adding Textblob Sentiment Features

body_dicts = []

for text in list(df['body']):
    col_dict = {}
    pol, sub = TextBlob(text).sentiment
    col_dict['polarity_bod'] = pol
    col_dict['subjectivity_bod'] = sub
    body_dicts.append(col_dict)
    
title_dicts = []

for text in list(df['title']):
    col_dict = {}
    pol, sub = TextBlob(text).sentiment
    col_dict['polarity_tit'] = pol
    col_dict['subjectivity_tit'] = sub
    title_dicts.append(col_dict)

df = df.merge(pd.DataFrame(title_dicts), left_index=True, right_index=True)
df = df.merge(pd.DataFrame(body_dicts), left_index=True, right_index=True)
df.shape

## Further Exploration

# popular alt-right symbol for pepe, doesn't show up in any r/Republican posts

# df[df['title'].map(lambda x: True if '🐸' in x else False)]

# df[df['title'].map(lambda x: True if 'immigrant' in x.lower() else False)]

# cf.set_config_file(offline=False, world_readable=True, theme='ggplot')

# df.groupby('is_the_donald').mean().T.iplot(kind='bar',
#                yTitle='Analysis score',
#                title='Mean Sentment Analysis by Subreddit',
#                filename='Average_Sentiments')

# 9 of top 10 lowest compound titles were the_Donald

df.sort_values('vad_title_compound').head(10)

# different story for bodies

## Export the Data to CSV 

# export to csv for manipulation
# df.to_csv('data/final_cleaned.csv', index=False)
