==> 03_Modeling.ipynb <==
# Project 3: Reddit Post Classification

<i>Pulling information and classifying posts via Pushshift's API</i>

**Author: Brendan McDonnell**

## Step 3: Modeling

Creating a model that predicts which reddit a post belongs in.

## Relative Links
- [Importing Libraries and Datasets Needed](#Importing-Libraries-and-Datasets-Needed)
- [Deciding Between TFIDF & CVEC](#Deciding-Between-TFIDF-&-CVEC)
- [Checking Sentiment Analysis Features](#Checking-Sentiment-Analysis-Features)
- [Modeling Predictions](#Modeling-Predictions)

## Importing Libraries and Datasets Needed

import pandas as pd
import numpy as np

from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.pipeline import Pipeline
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.linear_model import LogisticRegression

df = pd.read_csv('./data/final_cleaned.csv')

df.head()

# baseline
df.is_the_donald.value_counts(normalize=True)

## Deciding Between TFIDF & CVEC

Because we need to feed an X series feature, I will do run each vectorizer in a pipeline for both the 'title' and 'body' columns w/ a basic logistic regression. Will decided between the methods after seeing which works best with this dataset.

X_t = df['title'] # X title series
X_b = df['body'] # X body series
X_sent = df[['vad_title_neg',
             'vad_title_neu',
             'vad_title_pos', 
             'vad_title_compound', 
             'vad_body_neg',
             'vad_body_neu',
             'vad_body_pos',
             'vad_body_compound',
             'polarity_tit',
             'subjectivity_tit',
             'polarity_bod',
             'subjectivity_bod']]
y = df['is_the_donald']

# train test split the two X's y's for titles and bodies
X_train_t, X_test_t, y_train_t, y_test_t = train_test_split(X_t, y, stratify=y, random_state=4)
X_train_b, X_test_b, y_train_b, y_test_b = train_test_split(X_b, y, stratify=y, random_state=4)
X_train_sent, X_test_sent, y_train, y_test = train_test_split(X_sent, y, stratify=y, random_state=4)

# two pipelines
pipe_cvec = Pipeline([('cvec', CountVectorizer()),
             ('lr', LogisticRegression())])
pipe_tvec = Pipeline([('tvec', TfidfVectorizer()),
             ('lr', LogisticRegression())])

X_train_t.shape, X_test_t.shape

X_train_b.shape, X_test_b.shape

X_train_sent.shape, X_test_sent.shape

# pipe params for CountVectorizer
pipe_cvec_params = {
    'cvec__max_features': [2500, 3000, 3500],
    'cvec__min_df': [2, 3],
    'cvec__max_df': [.9, .95],
    'cvec__stop_words': [None, 'english'],
    'cvec__ngram_range': [(1,1), (1,2), (1,3)]
}

# grid search over title training data for CVEC
gs_cvec_t = GridSearchCV(pipe_cvec, param_grid=pipe_cvec_params, cv=3, verbose=1)
gs_cvec_t.fit(X_train_t, y_train_t)
print(gs_cvec_t.best_score_)
gs_cvec_t.best_params_

# pipe params for CountVectorizer
pipe_cvec_params = {
    'cvec__max_features': [3500],
    'cvec__min_df': [2],
    'cvec__max_df': [.9],
    'cvec__stop_words': ['english'],
    'cvec__ngram_range': [(1,2)]
}

# grid search over body training data for CVEC
gs_cvec_b = GridSearchCV(pipe_cvec, param_grid=pipe_cvec_params, cv=3, verbose=1)
gs_cvec_b.fit(X_train_b, y_train_b)
print(gs_cvec_b.best_score_)
gs_cvec_b.best_params_

# pipe params for TFIDF Vectorizer
pipe_tvec_params = {
    'tvec__max_features': [10000, 9500],
    'tvec__min_df': [3],
    'tvec__max_df': [.9],
    'tvec__stop_words': [None],
    'tvec__ngram_range': [(1,1)]
}

# grid search over title training data for TFID
gs_tvec_t = GridSearchCV(pipe_tvec, param_grid=pipe_tvec_params, cv=3, verbose=1)
gs_tvec_t.fit(X_train_t, y_train_t)
print(gs_tvec_t.best_score_)
gs_tvec_t.best_params_

# pipe params for TFIDF Vectorizer
pipe_tvec_params = {
    'tvec__max_features': [2500, 3000, 3500],
    'tvec__min_df': [2, 3],
    'tvec__max_df': [.9, .95],
    'tvec__stop_words': [None, 'english'],
    'tvec__ngram_range': [(1,1), (1,2), (1,3)]
}

# grid search over body training data for TFID
gs_tvec_b = GridSearchCV(pipe_tvec, param_grid=pipe_tvec_params, cv=3, verbose=1)
gs_tvec_b.fit(X_train_b, y_train_b)
print(gs_tvec_b.best_score_)
gs_tvec_b.best_params_

gs_cvec_t.score(X_test_t, y_test_t), gs_cvec_b.score(X_test_b, y_test_b), gs_tvec_t.score(X_test_t, y_test_t), gs_tvec_b.score(X_test_b, y_test_b)

TFIDF Vectorizer seems to be the marginally better option for vectorizing the words in the title. Removing stop words and using CVEC is better for the body (which isn't surprising)... The predictions will never be that good for the bodies given that about 1/3 of the datum actually contain any body at all.

# after a bit of testing;
gs_tvec_t.best_params_, gs_cvec_b.best_params_

## Checking Sentiment Analysis Features

**Are these features worth including in the model? Computational efficiency played a big part in me being allowed to include these.**

X_sent = df[['vad_title_neg', 'vad_title_neu',
       'vad_title_pos', 'vad_title_compound', 'vad_body_neg', 'vad_body_neu',
       'vad_body_pos', 'vad_body_compound', 'polarity_tit', 'subjectivity_tit',
       'polarity_bod', 'subjectivity_bod']]
y = df['is_the_donald']

X_train, X_test, y_train, y_test = train_test_split(X_sent, y, stratify=y, random_state=4)

logreg = LogisticRegression()

logreg.fit(X_train, y_train)
logreg.score(X_train, y_train)

**Answer: No**

## Modeling Predictions

**Fitting a Gaussian Naive Bayes model results in an overfit model, performing .10 above baseline; in fact, just using TFIDF and logistic regression on the Title information results in a much better model, as shown above.**

### Logistic Regression

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.pipeline import Pipeline
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.linear_model import LogisticRegression
from sklearn.naive_bayes import GaussianNB
from sklearn.ensemble import RandomForestClassifier

X = df['title'] + ' ' + df['body']
y = df['is_the_donald']

X_train, X_test, y_train, y_test = train_test_split(X, y, stratify=y, random_state=4)

# baseline
1 - y.mean()

# logistic regression
params = [{'tvec__max_df': [0.9],
           'tvec__max_features': [10_000],
           'tvec__min_df': [3],
           'tvec__ngram_range': [(1, 1)],
           'tvec__stop_words': [None],
           'lr__penalty': ['l2'],
           'lr__C': [1.5]
          }]
pipe_lr = Pipeline([('tvec', TfidfVectorizer()),
                     ('lr', LogisticRegression())
                    ])
grid = GridSearchCV(param_grid=params, estimator=pipe_lr, cv=3, verbose=1, n_jobs=2)
grid.fit(X_train, y_train)
print(grid.best_score_)
grid.best_params_

tvec = TfidfVectorizer(max_df = 0.9, max_features = 10_000, min_df = 3, ngram_range = (1,1), stop_words=None)
X_train = pd.DataFrame(tvec.fit_transform(X_train).todense(),
             columns=tvec.get_feature_names())
X_test  = pd.DataFrame(tvec.transform(X_test).todense(),
             columns=tvec.get_feature_names())
lr = LogisticRegression(C=1.5, penalty='l2')
lr.fit(X_train, y_train)

df_beta = pd.DataFrame(np.exp(lr.coef_), columns=X_train.columns)

df_beta.shape

# getting a visual of the beta values and which feature they are associated with.
df_beta.T.sort_values(by=0, ascending=False).head()

grid.score(X_test, y_test)
